auth/requirements.txt
fastapi[all]
fastapi-users[sqlalchemy]
pyjwt
passlib[bcrypt]
python-dotenv
psycopg2-binary
pytz